requirements.txt
orjson>=3.10.0
pydantic>=2.7.0
jsonschema>=4.22.0
pyyaml>=6.0.1
typer>=0.12.3
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
duckdb>=1.0.0
httpx>=0.27.0
rich>=13.7.1
pytest>=8.3.0
